requirements.txt
# AWS SDK asynchrone
aioboto3==12.3.0
botocore>=1.31.0

# PostgreSQL asynchrone (très performant)
asyncpg==0.29.0

# Configuration et validation
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logs structurés
structlog==24.1.0

# Métriques OpenTelemetry (export stdout)
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0

# Typing extensions
typing-extensions>=4.8.0
